+++ Generics/Generics/Program.cs
@@ -11,15 +11,15 @@ namespace Generics
         static void Main(string[] args)
         {
             Console.WriteLine("Please enter the first number");
-            int num1 = Convert.ToInt32(Console.ReadLine());
+            float num1 = Convert.ToInt32(Console.ReadLine());
 
             Console.WriteLine("Please enter the second number");
-            int num2 = Convert.ToInt32(Console.ReadLine());
+            float num2 = Convert.ToInt32(Console.ReadLine());
 
             Console.WriteLine("Please enter the third number");
-            int num3 = Convert.ToInt32(Console.ReadLine());
+            float num3 = Convert.ToInt32(Console.ReadLine());
 
-            int val = FindeMaxValue.MaximumIntegerNumber(num1, num2, num3);
+            float val = FindeMaxValue.MaximumIntegerNumber(num1, num2, num3);
             Console.WriteLine("{0} IS GREATER  NUMBER.", val);
         }
     }
